requirements.txt
discord.py
google-generativeai
yt-dlp
spotipy
aiohttp
aiosqlite
beautifulsoup4
youtube-transcript-api
requests
